# micro/openapi.yaml
...
openapi: 3.0.3
info:
    title: ""
    version: 0.0.1
paths:
...
    /v1/dict/data/delete/{id}:
        delete:
            tags:
                - SysDict
            operationId: SysDict_DeleteSysDictData
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.system.v1.EmptySysDictReply'
    /v1/dict/data/info/{id}:
        get:
            tags:
                - SysDict
            operationId: SysDict_GetSysDictData
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.system.v1.GetSysDictDataReply'
    /v1/dict/data/list:
        post:
            tags:
                - SysDict
            operationId: SysDict_ListSysDictData
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.system.v1.ListSysDictDataRep'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.system.v1.ListSysDictDataReply'
    /v1/dict/data/save:
        post:
            tags:
                - SysDict
            operationId: SysDict_CreateSysDictData
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.system.v1.SysDictDataRep'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.system.v1.EmptySysDictReply'
    /v1/dict/data/update:
        put:
            tags:
                - SysDict
            operationId: SysDict_UpdateSysDictData
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.system.v1.SysDictDataRep'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.system.v1.EmptySysDictReply'
    /v1/dict/data/{type}:
        get:
            tags:
                - SysDict
            operationId: SysDict_SysDictDataType
            parameters:
                - name: type
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.system.v1.SysDictDataTypeReply'
    /v1/dict/type/delete/{id}:
        delete:
            tags:
                - SysDict
            operationId: SysDict_DeleteSysDictType
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.system.v1.EmptySysDictReply'
    /v1/dict/type/info/{id}:
        get:
            tags:
                - SysDict
            operationId: SysDict_GetSysDictType
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.system.v1.GetSysDictTypeReply'
    /v1/dict/type/list:
        post:
            tags:
                - SysDict
            operationId: SysDict_ListSysDictType
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.system.v1.ListSysDictTypeRep'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.system.v1.ListSysDictTypeReply'
...
    /v1/dict/type/save:
        post:
            tags:
                - SysDict
            operationId: SysDict_CreateSysDictType
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.system.v1.SysDictTypeRep'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.system.v1.EmptySysDictReply'
    /v1/dict/type/update:
        put:
            tags:
                - SysDict
            operationId: SysDict_UpdateSysDictType
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.system.v1.SysDictTypeRep'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.system.v1.EmptySysDictReply'
...
components:
    schemas:
...
        api.system.v1.DictDataReply:
            type: object
            properties:
                dictCode:
                    type: string
                dictSort:
                    type: string
                dictLabel:
                    type: string
                dictValue:
                    type: string
                dictType:
                    type: string
                cssClass:
                    type: string
                listClass:
                    type: string
                isDefault:
                    type: string
                status:
                    type: string
                default:
                    type: boolean
                createdAt:
                    type: string
                remark:
                    type: string
        api.system.v1.DictTypeReply:
            type: object
            properties:
                dictId:
                    type: string
                createdAt:
                    type: string
                remark:
                    type: string
                dictName:
                    type: string
                dictType:
                    type: string
                status:
                    type: string
...
        api.system.v1.EmptySysDictReply:
            type: object
            properties: {}
...
        api.system.v1.GetSysDictDataReply:
            type: object
            properties:
                dict:
                    $ref: '#/components/schemas/api.system.v1.DictDataReply'
        api.system.v1.GetSysDictTypeReply:
            type: object
            properties:
                dict:
                    $ref: '#/components/schemas/api.system.v1.DictTypeReply'
...
        api.system.v1.ListSysDictDataRep:
            type: object
            properties:
                pageNum:
                    type: string
                    description: pageNum
                pageSize:
                    type: string
                    description: 页面大小
                dictType:
                    type: string
        api.system.v1.ListSysDictDataReply:
            type: object
            properties:
                total:
                    type: string
                rows:
                    type: array
                    items:
                        $ref: '#/components/schemas/api.system.v1.DictDataReply'
        api.system.v1.ListSysDictTypeRep:
            type: object
            properties:
                pageNum:
                    type: string
                    description: pageNum
                pageSize:
                    type: string
                    description: 页面大小
        api.system.v1.ListSysDictTypeReply:
            type: object
            properties:
                total:
                    type: string
                rows:
                    type: array
                    items:
                        $ref: '#/components/schemas/api.system.v1.DictTypeReply'
...
        api.system.v1.SysDictDataRep:
            type: object
            properties:
                dictCode:
                    type: string
                dictSort:
                    type: string
                dictLabel:
                    type: string
                dictValue:
                    type: string
                dictType:
                    type: string
                cssClass:
                    type: string
                listClass:
                    type: string
                isDefault:
                    type: string
                status:
                    type: string
                default:
                    type: boolean
                createAt:
                    type: string
                remark:
                    type: string
        api.system.v1.SysDictDataTypeReply:
            type: object
            properties:
                dict:
                    type: array
                    items:
                        $ref: '#/components/schemas/api.system.v1.DictDataReply'
        api.system.v1.SysDictTypeRep:
            type: object
            properties:
                dictId:
                    type: string
                dictName:
                    type: string
                dictType:
                    type: string
                status:
                    type: string
                remark:
                    type: string
...